# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Aanya - AI Companion API for breakup recovery
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OnboardingStatus {
  isComplete: boolean;
  /** name | relationshipType | energy | companionName | done */
  currentStep: string;
  /**
     * The companion's warm first greeting after onboarding completes
     * @nullable
     */
  companionFirstMessage?: string | null;
}

export interface OnboardingAnswerInput {
  /** name | relationshipType | energy | companionName */
  step: string;
  answer: string;
}

export interface Profile {
  id: number;
  userName: string;
  companionName: string;
  /** friend | romantic */
  relationshipType: string;
  /** playful | calm | deep */
  energy: string;
  createdAt: string;
  daysSinceStart: number;
  /** 1=Arrival 2=Held 3=FirstStep 4=Building */
  currentStage: number;
}

export interface ProfileInput {
  userName?: string;
  companionName?: string;
  relationshipType?: string;
  energy?: string;
}

export interface Message {
  id: number;
  /** user | assistant */
  role: string;
  content: string;
  createdAt: string;
  isMorningNote: boolean;
}

export interface MessageInput {
  /** @minLength 1 */
  content: string;
}

export interface ChatReply {
  userMessage: Message;
  assistantMessage: Message;
  /** Whether memory extraction ran this cycle */
  memoryExtracted: boolean;
}

export interface MemoryFact {
  id: number;
  fact: string;
  /** life | preference | event | person | goal */
  category: string;
  createdAt: string;
}

export interface PersonalitySignal {
  id: number;
  signal: string;
  observedCount: number;
  /** True after observed 3+ times */
  isActive: boolean;
  createdAt: string;
}

export interface Win {
  id: number;
  content: string;
  createdAt: string;
}

export interface WinInput {
  /** @minLength 1 */
  content: string;
}

export interface Habit {
  id: number;
  name: string;
  /** When-then anchor e.g. 'after morning chai' */
  whenThen: string;
  reason: string;
  isActive: boolean;
  streak: number;
  /** @nullable */
  lastCompleted?: string | null;
  /** Last 7 days of completion dates */
  recentCompletions: string[];
  createdAt: string;
}

export interface HabitInput {
  /** @minLength 1 */
  name: string;
  whenThen: string;
  reason: string;
}

export interface HabitUpdate {
  name?: string;
  whenThen?: string;
  reason?: string;
  isActive?: boolean;
}

export interface Milestone {
  id: string;
  label: string;
  isUnlocked: boolean;
  /** @nullable */
  unlockedAt?: string | null;
}

export interface Journey {
  /** 1=Arrival 2=Held 3=FirstStep 4=Building */
  stage: number;
  /** Arrival | Held | First Step | Building */
  stageLabel: string;
  dayCounter: number;
  streak: number;
  winCount: number;
  habitCount: number;
  /** @nullable */
  averageMoodRecent?: number | null;
  milestones: Milestone[];
  /** @nullable */
  moodCaption?: string | null;
}

export interface MoodEntry {
  date: string;
  score: number;
  createdAt: string;
}