# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Aanya - AI Companion API for breakup recovery
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /onboarding/answer:
    post:
      operationId: submitOnboardingAnswer
      tags: [onboarding]
      summary: Submit an onboarding answer (one step at a time)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingAnswerInput"
      responses:
        "200":
          description: Next onboarding step or completion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingStatus"
...
  /memory/wins:
    get:
      operationId: getWins
      tags: [memory]
      summary: Get all real-world wins logged
      responses:
        "200":
          description: List of wins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Win"
    post:
      operationId: createWin
      tags: [memory]
      summary: Log a new win
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WinInput"
      responses:
        "201":
          description: Created win
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Win"
  /memory/habits:
    get:
      operationId: getHabits
      tags: [memory]
      summary: Get all habits with streaks and recent completions
      responses:
        "200":
          description: List of habits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Habit"
    post:
      operationId: createHabit
      tags: [memory]
      summary: Create a new habit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HabitInput"
      responses:
        "201":
          description: Created habit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Habit"
  /memory/habits/{id}:
    put:
      operationId: updateHabit
      tags: [memory]
      summary: Update a habit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HabitUpdate"
      responses:
        "200":
          description: Updated habit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Habit"
  /memory/habits/{id}/complete:
    post:
      operationId: completeHabit
      tags: [memory]
      summary: Mark a habit as completed for today
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated habit with new streak
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Habit"
...
components:
  schemas:
...
    OnboardingStatus:
      type: object
      properties:
        isComplete:
          type: boolean
        currentStep:
          type: string
          description: "name | relationshipType | energy | companionName | done"
        companionFirstMessage:
          type: ["string", "null"]
          description: The companion's warm first greeting after onboarding completes
      required:
        - isComplete
        - currentStep
    OnboardingAnswerInput:
      type: object
      properties:
        step:
          type: string
          description: "name | relationshipType | energy | companionName"
        answer:
          type: string
      required:
        - step
        - answer
...
    Win:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - content
        - createdAt
    WinInput:
      type: object
      properties:
        content:
          type: string
          minLength: 1
      required:
        - content
    Habit:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        whenThen:
          type: string
          description: "When-then anchor e.g. 'after morning chai'"
        reason:
          type: string
        isActive:
          type: boolean
        streak:
          type: integer
        lastCompleted:
          type: ["string", "null"]
          format: date
        recentCompletions:
          type: array
          items:
            type: string
            format: date
          description: Last 7 days of completion dates
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - whenThen
        - reason
        - isActive
        - streak
        - createdAt
        - recentCompletions
    HabitInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        whenThen:
          type: string
        reason:
          type: string
      required:
        - name
        - whenThen
        - reason
    HabitUpdate:
      type: object
      properties:
        name:
          type: string
        whenThen:
          type: string
        reason:
          type: string
        isActive:
          type: boolean
...