# lib/api-zod/src/generated/types/habit.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Aanya - AI Companion API for breakup recovery
 * OpenAPI spec version: 0.1.0
 */

export interface Habit {
  id: number;
  name: string;
  /** When-then anchor e.g. 'after morning chai' */
  whenThen: string;
  reason: string;
  isActive: boolean;
  streak: number;
  /** @nullable */
  lastCompleted?: Date | null;
  /** Last 7 days of completion dates */
  recentCompletions: Date[];
  createdAt: Date;
}