# lib/api-zod/src/generated/types/habitUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Aanya - AI Companion API for breakup recovery
 * OpenAPI spec version: 0.1.0
 */

export interface HabitUpdate {
  name?: string;
  whenThen?: string;
  reason?: string;
  isActive?: boolean;
}