# lib/api-zod/src/generated/types/onboardingStatus.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Aanya - AI Companion API for breakup recovery
 * OpenAPI spec version: 0.1.0
 */

export interface OnboardingStatus {
  isComplete: boolean;
  /** name | relationshipType | energy | companionName | done */
  currentStep: string;
  /**
     * The companion's warm first greeting after onboarding completes
     * @nullable
     */
  companionFirstMessage?: string | null;
}