# lib/api-zod/src/generated/types/winInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Aanya - AI Companion API for breakup recovery
 * OpenAPI spec version: 0.1.0
 */

export interface WinInput {
  /** @minLength 1 */
  content: string;
}